Switch client/state/http from superagent to fetch.
This handler was currently using `superagent`, a third-party library for
performing network requests. This library precedes `fetch` and nowadays
doesn't offer much in comparison, other than slightly easier error
handling.

For both security and bundle size reasons, it's best to move away from
it to the natively-supported (and elsewhere polyfilled) `fetch`.

This is the first of several PRs towards that effort.